_something_ is being displayed

+++ include/shaders/Shader.h
@@ -12,10 +12,10 @@ public:
   virtual ~Shader() = default;
   virtual void run(VertexArrayObject* inStream, VertexArrayObject* outStream, std::vector<int> attrIndices) = 0;
 
-  const std::map<int, int>& getInputBufferMap() const; 
-  const std::map<int, int>& getOutputBufferMap() const; 
+  const std::map<size_t, size_t>& getInputBufferMap() const; 
+  const std::map<size_t, size_t>& getOutputBufferMap() const; 
 protected:
   // attributeIndex -> itemSize (everything is single-buffer)
-  std::map<int, int> inputBufferMap;
-  std::map<int, int> outputBufferMap;
+  std::map<size_t, size_t> inputBufferMap;
+  std::map<size_t, size_t> outputBufferMap;
 };
