+++ src/shaders/Shader.cc
@@ -1,10 +1,10 @@
 #include "VertexArrayObject.h"
 #include "shaders/Shader.h"
 
-const std::map<int, int>& Shader::getInputBufferMap() const {
+const std::map<size_t, size_t>& Shader::getInputBufferMap() const {
   return inputBufferMap;
 }
 
-const std::map<int, int>& Shader::getOutputBufferMap() const {
+const std::map<size_t, size_t>& Shader::getOutputBufferMap() const {
   return outputBufferMap;
 }
